♻️ [Refactor] dataset, apply tensorlize
